introduce useRoom, refactor
