+++ src/contexts/WargameProvider.tsx
@@ -1,6 +1,7 @@
 import { useState, ReactNode } from 'react'
-import { ForceDetails, GameStateType, RoomType, UserDetailsType } from '../types/wargame'
+import { ForceDetails, GameStateType, UserDetailsType } from '../types/wargame'
 import { WargameContext } from './WargameContext'
+import { RoomType } from '../types/rooms'
 
 interface WargameProviderProps {
   children: ReactNode
@@ -11,7 +12,8 @@ export const WargameProvider = ({ children }: WargameProviderProps) => {
   const [gameState, setGameState] = useState<GameStateType | null>(null)
   const [rooms, setRooms] = useState<RoomType[]>([])
   const [playerForce, setPlayerForce] = useState<ForceDetails | null>(null)
-  const [playerDetails, setPlayerDetails] = useState<UserDetailsType | null>(null)    
+  const [playerDetails, setPlayerDetails] = useState<UserDetailsType | null>(null)   
+  const [useMock, setUseMock] = useState(false) 
 
   const value = {
     loggedIn,
@@ -23,7 +25,9 @@ export const WargameProvider = ({ children }: WargameProviderProps) => {
     playerForce,
     setPlayerForce,
     playerDetails,
-    setPlayerDetails
+    setPlayerDetails,
+    useMock,
+    setUseMock
   }
 
   return (
